Moved the exports to the index file in the qp
