+++ src/AppEntry.js
@@ -4,8 +4,8 @@ import { Provider } from 'react-redux';
 import { init } from './store';
 import App from './App';
 import getBaseName from './Utilities/getBaseName';
-import FeatureFlagProvider from './FeatureFlags/Provider';
-import QueryParamsProvider from './QueryParams/Provider';
+import { FeatureFlagProvider } from './FeatureFlags';
+import { QueryParamsProvider } from './QueryParams';
 
 const AutomationAnalytics = () => (
   <Provider store={init().getStore()}>
