+++ src/Containers/JobExplorer/JobExplorer.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useCallback } from 'react';
 
-import { useQueryParams } from '../../QueryParams/useQueryParams';
+import { useQueryParams } from '../../QueryParams/';
 import useRequest from '../../Utilities/useRequest';
 
 import LoadingState from '../../Components/ApiStatus/LoadingState';
