+++ src/FeatureFlags/index.ts
@@ -1,3 +1,3 @@
 export { ValidFeatureFlags } from './types';
 export { default as useFeatureFlag } from './useFeatureFlag';
-export { default as Provider } from './Provider';
+export { default as FeatureFlagProvider } from './Provider';
