+++ src/QueryParams/index.ts
@@ -1,0 +1,2 @@
+export { default as useQueryParams } from './useQueryParams';
+export { default as QueryParamsProvider } from './Provider';
